16/8/2566

+++ myApp/Models/Remark.cs
@@ -1,13 +1,14 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Numerics;
 using System.Web;
 
 namespace myApp.Models
 {
     public class RemarkHS
     {
-        public int Id { get; set; }
+        public BigInteger Id { get; set; }
         public string FormGuid { get; set; }
         public string UserK2 { get; set; }
         public string Name { get; set; }
